requirements.txt
# flake8
flake8==3.7.7
pyflakes==2.1.1

# from sqlsoup
wheel==0.32.3
sqlalchemy==1.1.9

# from bs4
beautifulsoup4==4.6.0

# from pytest
setuptools==34.0.2
py==1.5.2
funcsigs==1.0.2

# from pytest, setuptools, allure-pytest
more-itertools==4.3.0
six==1.10.0

# from pytest-xdist
apipkg==1.4
execnet==1.4.1
pytest-forked==0.2

# from pytest-sugar
termcolor==1.1.0

# from allure-pytest
allure-python-commons==2.5.1

# from allure-pytest, allure-python-commons
attrs==18.2.0
pluggy==0.6.0

# for autotests
selenium==3.14.0
sqlsoup==0.9.1
bs4==0.0.1
fake-factory==0.5.3
pytest==3.6
pytest-xdist==1.23.0
pytest_rerunfailures==4.1
pytest-instafail==0.3.0
allure-pytest==2.5.1
numpy==1.15.0
requests==2.10.0
Pillow==5.3.0